+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## X.Y.Z (INSERT_DATE_HERE)
+## 0.7.1 (4-3-2016)
 
 #### Minor Changes
 - [Parsing Errors should include Parser context information.](#165)